Rename common.py to const.py
